refactor/test : fake 객체를 따로 패키지를 관리 및 가격 변동 테스트 추가
