+++ src/test/java/kitchenpos/products/tobe/fakes/FakePurgomalums.java
@@ -1,4 +1,6 @@
-package kitchenpos.products.tobe.domain;
+package kitchenpos.products.tobe.fakes;
+
+import kitchenpos.products.tobe.domain.ProfanitiesChecker;
 
 import java.util.Arrays;
 import java.util.List;
